CreateNewAccountView
Added a form for adding a new account (money account). In the form you can specify the name, amount of funds and color of the card. There is also a preview at the top so you can see what the new account will look like. Also added logic to open this form in sheets